fix:refresh auth token

+++ src/app/interceptors/http-interceptor.service.ts
@@ -10,34 +10,53 @@ import { Router } from '@angular/router';
 })
 export class HttpInterceptorService {
   constructor(private authService: AuthService, private snackbar: MatSnackBar, private router: Router) { }
+  maxRetry = 3;
+  retryCount = 0;
   intercept(req: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
     const accessToken = sessionStorage.getItem('access_token');
     const sessionState:any = sessionStorage.getItem('session_state');
 
     if (accessToken && sessionState) {
-      req = this.addToken(req, accessToken, sessionState);
+      req = this.addToken(req);
     }
 
     return next.handle(req).pipe(
       catchError((error: HttpErrorResponse) => {
-        if (error.error?.error?.includes('invalid_grant')) {
-          this.snackbar.open('Session expired, Login again', 'Ok', {
-            horizontalPosition: 'right',
-            verticalPosition: 'top',
-            duration: 10000,
-            panelClass: 'app-notification-error',
-          });
-            this.router.navigate(['../']);
-            sessionStorage.clear();
-          return EMPTY;
+          if(error.url?.includes('refresh-token')){
+            if(this.maxRetry <= this.retryCount){
+              this.retryCount++;
+              return EMPTY;
+            }
+            else if(error.error?.error?.includes('invalid_grant')){
+              this.snackbar.open('Session expired, Login again', 'Ok', {
+                horizontalPosition: 'right',
+                verticalPosition: 'top',
+                duration: 10000,
+                panelClass: 'app-notification-error',
+              });
+              this.router.navigate(['../']);
+              sessionStorage.clear();
+              return EMPTY;
+            }
+            else{
+              this.snackbar.open('Refresh token expired, Please login again', 'Ok', {
+                horizontalPosition: 'right',
+                verticalPosition: 'top',
+                duration: 10000,
+                panelClass: 'app-notification-error',
+              });
+              this.router.navigate(['../']);
+              sessionStorage.clear();
+              return EMPTY;
+            }
         }
-        else{
-          if (!error.error.status && accessToken) {
+        else {
+          if (!error.error.status && accessToken && sessionState) {
               return this.authService.refreshToken().pipe(
                 switchMap((refreshTokenDetails: any) => {
                   sessionStorage.setItem('access_token', refreshTokenDetails.accessToken);
                   sessionStorage.setItem('session_state', refreshTokenDetails.sessionState);
-                  req = this.addToken(req, refreshTokenDetails.accessToken, refreshTokenDetails.sessionState);
+                  req = this.addToken(req);
                   return next.handle(req);
                 }),
                 catchError((refreshError: any) => {
@@ -52,7 +71,7 @@ export class HttpInterceptorService {
     );
   }
 
-  private addToken(req: HttpRequest<any>, accessToken: string, sessionState: string): HttpRequest<any> {
+  private addToken(req: HttpRequest<any>): HttpRequest<any> {
     return req.clone({
       setHeaders: {
         Authorization: `${sessionStorage.getItem('access_token')}`,
